Simplified EventLog utility name from MultitenantEventLog
